Added the passlib[bcrypt] library.
